requirements.txt
flask
gunicorn
line-bot-sdk
pythainlp
pandas
